Add files via upload
Design and Implementation of a movie booking system. With the requirements as shared in the Google Doc.

Tried using Facade Pattern to trigger flow of information.
Once user registers and select an option of creating a movie, he is redirected to select his Movie Timings and Seat Allocation. The application then creates an unique booking ID and the booking details and then passes the confirmation to the User. If not, the user is prompted to input the data again.